Move fileset expansion to project loading
