+++ snapcraft/internal/project_loader.py
@@ -32,7 +32,7 @@ from snapcraft.internal import (
     pluginhandler,
 )
 from snapcraft._schema import Validator, SnapcraftSchemaError
-from snapcraft.internal.parts import get_remote_parts
+from snapcraft.internal.parts import SnapcraftLogicError, get_remote_parts
 
 
 logger = logging.getLogger(__name__)
@@ -127,6 +127,7 @@ class Config:
         self._validator.validate()
 
         snapcraft_yaml = self._process_remote_parts(snapcraft_yaml)
+        snapcraft_yaml = self._expand_filesets(snapcraft_yaml)
         self.data = self._expand_env(snapcraft_yaml)
 
         # both confinement type and build quality are optionals
@@ -207,6 +208,16 @@ class Config:
                     ('$SNAPCRAFT_PROJECT_VERSION', snapcraft_yaml['version']),
                     ('$SNAPCRAFT_STAGE', self._project_options.stage_dir),
                 ])
+        return snapcraft_yaml
+
+    def _expand_filesets(self, snapcraft_yaml):
+        parts = snapcraft_yaml.get('parts', {})
+
+        for part_name in parts:
+            for step in ('stage', 'snap'):
+                step_fileset = _expand_filesets_for(step, parts[part_name])
+                parts[part_name][step] = step_fileset
+
         return snapcraft_yaml
 
     def _process_remote_parts(self, snapcraft_yaml):
@@ -379,3 +390,22 @@ def _ensure_grade_default(yaml_data, schema):
         logger.warning('"grade" property not specified: defaulting '
                        'to "stable"')
         yaml_data['grade'] = schema['grade']['default']
+
+
+def _expand_filesets_for(step, properties):
+    filesets = properties.get('filesets', {})
+    fileset_for_step = properties.get(step, {})
+    new_step_set = []
+
+    for item in fileset_for_step:
+        if item.startswith('$'):
+            try:
+                new_step_set.extend(filesets[item[1:]])
+            except KeyError:
+                raise SnapcraftLogicError(
+                    '\'{}\' referred to in the \'{}\' fileset but it is not '
+                    'in filesets'.format(item, step))
+        else:
+            new_step_set.append(item)
+
+    return new_step_set
